v2: better interface and methods implementation (#18)
* credentials: new interface to access with the credentials

* configuration: methods for set the username

* configuration: changing from public to private struct members

* configuration: refactoring and call the credentials methods, that avoid to call the function twice for the same action

* credentials: handled when the credentials are not set

* examples: examples updated

+++ v2/credentials/credentials.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"io/ioutil"
 	"os"
+	"strings"
 )
 
 var credentials *Credentials
@@ -20,17 +21,35 @@ func New() *Credentials {
 	return credentials
 }
 
-func (c *Credentials) ReadFromFile(filename string) error {
+func (c *Credentials) SetFilename(filename string) {
 	c.credentialsFile = filename
+}
 
-	file, err := ioutil.ReadFile(filename)
+func (c *Credentials) GetFilename() string {
+	return c.credentialsFile
+}
+
+func (c *Credentials) ReadFromFile() error {
+	file, err := ioutil.ReadFile(c.credentialsFile)
 	if err != nil {
 		return err
 	}
 
-	if err := json.Unmarshal(file, &c); err != nil {
+	tmpCredentials := struct {
+		Id       string `json:"id"`
+		Password string `json:"password"`
+		Username string `json:"username"`
+		Email    string `json:"email"`
+	}{}
+
+	if err := json.Unmarshal(file, &tmpCredentials); err != nil {
 		return err
 	}
+
+	c.id = tmpCredentials.Id
+	c.password = tmpCredentials.Password
+	c.username = tmpCredentials.Username
+	c.email = tmpCredentials.Email
 
 	return nil
 }
@@ -40,7 +59,23 @@ func (c *Credentials) Save() error {
 		return errors.New("you must set the credential file")
 	}
 
-	file, err := json.MarshalIndent(c, "", " ")
+	if credentials == nil {
+		return errors.New("credentials cannot be blank")
+	}
+
+	tmpCredentials := struct {
+		Id       string `json:"id"`
+		Password string `json:"password"`
+		Username string `json:"username"`
+		Email    string `json:"email"`
+	}{
+		c.id,
+		c.password,
+		c.username,
+		c.email,
+	}
+
+	file, err := json.MarshalIndent(tmpCredentials, "", " ")
 	if err != nil {
 		return err
 	}
@@ -49,6 +84,10 @@ func (c *Credentials) Save() error {
 }
 
 func (c *Credentials) Remove() error {
+	if credentials == nil {
+		return errors.New("credentials not set")
+	}
+
 	return os.Remove(c.credentialsFile)
 }
 
@@ -57,4 +96,42 @@ func Get() Credentials {
 		return Credentials{}
 	}
 	return *credentials
+}
+
+func (c *Credentials) GetValue(key string) (string, error) {
+	key = strings.ToLower(key)
+	switch key {
+	case "id":
+		return c.id, nil
+	case "username":
+		return c.username, nil
+	case "email":
+		return c.email, nil
+	case "password":
+		return c.password, nil
+	default:
+		return "", errors.New("key does not exists")
+	}
+}
+
+func (c *Credentials) SetValue(key, value string) error {
+	key = strings.ToLower(key)
+	switch key {
+	case "id":
+		c.id = value
+	case "username":
+		c.username = value
+	case "email":
+		c.email = value
+	case "password":
+		c.password = value
+	default:
+		return errors.New("key does not exists")
+	}
+
+	return nil
+}
+
+func (c *Credentials) UnsetValue(key string) error {
+	return c.SetValue(key, "")
 }
